[Feat] order-service 코드 정리 및 FeignClient 뼈대 작성  (#63)
* fix : Jpa 구현체 분리,내부 컨트롤러 클래스명 변경

* feat: FeignClient 작성
- FeignClient 6개 작성 (Product/Company/Hub/User/Delivery/Notification)

* feat: FeignClient 응답 DTO 작성

* test: HTTP Client test 작성/실행

* fix: 코드래빗 권장 수정 사항 반영
- test.http 파일 위치 이동
- X-Internal-Request 헤더 검증 추가
- StockUpdateRequest DTO 추가 + Client 수정

* fix: unitPrice/totalPrice BigDecimal -> Integer 변경

+++ order-service/src/main/java/com/sparta/lucky/order/application/OrderService.java
@@ -15,7 +15,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Slf4j
@@ -37,7 +36,7 @@ public class OrderService {
      *
      * */
         String productName = "더미상품";
-        BigDecimal unitPrice = BigDecimal.valueOf(1000);
+        Integer unitPrice = 1000;
 
         Order order = Order.create(
                 request.getRequesterCompanyId(),
